add gatsby-node ts
